pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mythosmud"
version = "0.1.0"
description = "MythosMUD - A Cthulhu Mythos-themed MUD"
requires-python = ">=3.12"

dependencies = [
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "pydantic==2.11.7",
    "requests==2.32.4",
    "httpx==0.28.1",
    "argon2-cffi>=23.1.0",
    "python-jose==3.5.0",
    "faker>=37.5.3",
    "python-dotenv==1.0.0",
    "sqlalchemy>=2.0.42",
    "structlog>=25.4.0",
    "fastapi-users[sqlalchemy]>=14.0.0",
    "setuptools<81",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "aiosqlite>=0.21.0",
    "pytest-asyncio>=0.24.0",
    "jsonschema>=4.21.0",
    "bleach>=6.1.0",
    "tqdm>=4.67.1",
    "click>=8.1.0", # Command line interface for command parsing and validation
    "pwdlib[bcrypt]==0.2.1", # Required by fastapi-users (we override the hashing)
    "nats-py>=2.6.0", # NATS Python client for messaging
    "psutil>=6.0.0", # System and process utilities for memory monitoring
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-cov==6.2.1",
    "networkx>=3.5",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pytest-asyncio==0.24.0",
    "debugpy==1.8.15",
    "ruff==0.12.5",
    "aiosqlite>=0.21.0",
    "pre-commit>=4.2.0",
    "semgrep>=1.132.0",  # Static analysis for security and best practices
    "websocket-client>=1.8.0",  # WebSocket client for testing connection termination
]

[tool.setuptools.packages.find]
include = ["server*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by line-length
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
    "semgrep>=1.132.0",  # Static analysis for security and best practices
    "websocket-client>=1.8.0",  # WebSocket client for testing connection termination
]
